// Week 5/src/Type/Interfaces.tsx
export default interface Product {
    id: number;
    title: string;
    category: string;
    brand: string;
    warrantyInformation: string;
    description: string;
    price: number;
    images: string[];
  }

  export default interface FormData {
    name: string;
    email: string;
    password: string;
    phone: number; 
    gender: string;
    message: string;
    comment: string;
  }


  